chore(river_server): rename
